% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aws.transcribe-package}
\alias{aws.transcribe-package}
\alias{aws.transcribe}
\title{aws.transcribe}
\description{
A Cloudyr Project Package
}
\details{
This is a template package for the cloudyr project
}
\seealso{
\code{\link{start_transcription}}, \code{\link{list_transcriptions}}, \code{\link{get_transcription}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
